requirements.txt
pandas==1.3.5
python-telegram-bot==13.10